Adding the evaluation part in the inference script. Tried to use the tensorflow-addons package for weighted F1 score but couldn't figure out how to make it run.
It runs perfectly well when training and testing in the same script.
But if I save the model using model.save() and then try to read it using tf.keras.models.load_model(), it throws error.
Tried multiple things but couldn't figure it out. So chucking it for now.

Will not be using weighted F1-score while training but will use it at the end during evaluation using the sklearn library.
Maybe will get back at it again sometime later. But chucking for now.